ioctl cmd arg is u_long, strategy returns void.
SCCS-vsn: sys/hp300/dev/dcm.c 8.5
SCCS-vsn: sys/hp300/dev/sd.c 8.6
SCCS-vsn: sys/hp300/dev/ppi.c 8.2
SCCS-vsn: sys/hp300/dev/rd.c 8.3
SCCS-vsn: sys/hp300/dev/ct.c 8.3
SCCS-vsn: sys/hp300/dev/st.c 8.5

+++ usr/src/sys/hp300/dev/sd.c
@@ -7,7 +7,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)sd.c	8.5 (Berkeley) %G%
+ *	@(#)sd.c	8.6 (Berkeley) %G%
  */
 
 /*
@@ -996,7 +996,7 @@ sdwrite(dev, uio, flags)
 int
 sdioctl(dev, cmd, data, flag, p)
 	dev_t dev;
-	int cmd;
+	u_long cmd;
 	caddr_t data;
 	int flag;
 	struct proc *p;
